[stratumv2]: rename file connection_encryptor to encryptor
