+++ stratumv2/src/network/encryptor.rs
