+++ stratumv2/src/network/mod.rs
@@ -1,11 +1,11 @@
 mod channel;
 mod config;
-mod connection_encryptor;
+mod encryptor;
 mod message_handler;
 mod message_send_event;
 
 pub use channel::{new_channel_id, ChanID, Channel};
 pub use config::Config;
-pub use connection_encryptor::{ConnectionEncryptor, Encryptor};
+pub use encryptor::{ConnectionEncryptor, Encryptor};
 pub use message_handler::{JobNegotiationInitiator, MiningInitiator, NewConnReceiver};
 pub use message_send_event::MessageSendEvent;
